requirements.txt
google-analytics-data
google-auth
google-ads
pandas
matplotlib
jinja2
python-dotenv
